+++ src/lib/djerba/plugins/failed_report/test/plugin_test.py
@@ -23,7 +23,7 @@ class TestFailedReportPlugin(PluginTester):
         params = {
             self.INI: 'failed_report.ini',
             self.JSON: json_location,
-            self.MD5: '7f76c4a89ef6f1050004e11fbcce8ff6'
+            self.MD5: '7282fdbbc16e3230be229a40093e53d1'
         }
         self.run_basic_test(test_source_dir, params)